// src/common/types/enums.ts
export enum Action {
  add = "add",
  edit = "edit"
}

export enum Language {
  hy = "hy",
  ru = "ru",
  en = "en",
  "zh-CN" = "zh-CN"
}

export enum ProcessingTypes {
  sliced = "sliced",
  flattened = "flattened",
  antique = "antique",
  semi_gloss = "semi_gloss",
  gloss = "gloss",
  shub = "shub",
  brushatka = "brushatka"
}
export enum ThicknessTypes {
  fifteen = "fifteen",
  twenty = "twenty",
  thirty = "thirty"
}
export enum LengthTypes {
  four_hundred = "four_hundred",
  six_hundred = "six_hundred",
  eight_hundred = "eight_hundred",
  one_thousand_two_hundred = "one_thousand_two_hundred"
}
export enum WidthTypes {
  three_hundred = "three_hundred",
  four_hundred = "four_hundred"
}
export enum StoneTypes {
  travertine = "travertine",
  basalt = "basalt",
  tuff = "tuff",
  felsite = "felsite",
  granite = "granite"
}

export enum Currencies {
  USD = "USD",
  RUB = "RUB",
  CNY = "CNY"
}

export enum ApplicationStatus {
  reject = "reject",
  approve = "approve"
}